Language.Haskell.Tools.Refactor.Builtin.ExtensionOrganizer.ExtMap

Documentation

type ExtMap = Map (LogicalRelation Extension) [Occurence SrcSpan]

data Occurence a

Constructors

Hint

Fields

unOcc :: a

Evidence

Fields

unOcc :: a

MissingInformation

Fields

unOcc :: a

type LogicalRelation a = Formula a

prettyPrintFormula :: Show a => LogicalRelation a -> String

lVar :: a -> LogicalRelation a

lNot :: a -> LogicalRelation a

lAnd :: a -> a -> LogicalRelation a

lOr :: a -> a -> LogicalRelation a

lImplies :: a -> a -> LogicalRelation a

lAll :: [LogicalRelation a] -> LogicalRelation a

complexity :: Extension -> Int

expandExtension' :: Extension -> [Extension]

Completely expand extension

findFixedPoint :: Eq a => (a -> a) -> a -> a

Terminating variant of Control.Monad.Fix.fix It tries to find the fixpoint of a function with a given starting value. It terminates if a value repeats itself.

determineExtensions :: Map (LogicalRelation Extension) v -> [Extension]

rmImplied :: Extension -> [Extension] -> [Extension]

mergeImplied :: [Extension] -> [Extension]

data Extension

The language extensions known to GHC.

Note that there is an orphan Binary instance for this type supplied by the GHC.LanguageExtensions module provided by ghc-boot. We can't provide here as this would require adding transitive dependencies to the template-haskell package, which must have a minimal dependency set.

Constructors

Cpp

OverlappingInstances

UndecidableInstances

IncoherentInstances

UndecidableSuperClasses

MonomorphismRestriction

MonoPatBinds

MonoLocalBinds

RelaxedPolyRec

ExtendedDefaultRules

ForeignFunctionInterface

UnliftedFFITypes

InterruptibleFFI

CApiFFI

GHCForeignImportPrim

JavaScriptFFI

ParallelArrays

Arrows

TemplateHaskell

TemplateHaskellQuotes

QuasiQuotes

ImplicitParams

ImplicitPrelude

ScopedTypeVariables

AllowAmbiguousTypes

UnboxedTuples

UnboxedSums

BangPatterns

TypeFamilies

TypeFamilyDependencies

TypeInType

OverloadedStrings

OverloadedLists

NumDecimals

DisambiguateRecordFields

RecordWildCards

RecordPuns

ViewPatterns

GADTs

GADTSyntax

NPlusKPatterns

DoAndIfThenElse

BlockArguments

RebindableSyntax

ConstraintKinds

PolyKinds

DataKinds

InstanceSigs

ApplicativeDo

StandaloneDeriving

DeriveDataTypeable

AutoDeriveTypeable

DeriveFunctor

DeriveTraversable

DeriveFoldable

DeriveGeneric

DefaultSignatures

DeriveAnyClass

DeriveLift

DerivingStrategies

DerivingVia

TypeSynonymInstances

FlexibleContexts

FlexibleInstances

ConstrainedClassMethods

MultiParamTypeClasses

NullaryTypeClasses

FunctionalDependencies

UnicodeSyntax

ExistentialQuantification

MagicHash

EmptyDataDecls

KindSignatures

RoleAnnotations

ParallelListComp

TransformListComp

MonadComprehensions

GeneralizedNewtypeDeriving

RecursiveDo

PostfixOperators

TupleSections

PatternGuards

LiberalTypeSynonyms

RankNTypes

ImpredicativeTypes

TypeOperators

ExplicitNamespaces

PackageImports

ExplicitForAll

AlternativeLayoutRule

AlternativeLayoutRuleTransitional

DatatypeContexts

NondecreasingIndentation

RelaxedLayout

TraditionalRecordSyntax

LambdaCase

MultiWayIf

BinaryLiterals

NegativeLiterals

HexFloatLiterals

DuplicateRecordFields

OverloadedLabels

EmptyCase

PatternSynonyms

PartialTypeSignatures

NamedWildCards

StaticPointers

TypeApplications

Strict

StrictData

MonadFailDesugaring

EmptyDataDeriving

NumericUnderscores

QuantifiedConstraints

StarIsType

data Formula v

Constructors

Var v

Yes

No

Not (Formula v)

(Formula v) :&&: (Formula v)

(Formula v) :||: (Formula v)

(Formula v) :++: (Formula v)

(Formula v) :->: (Formula v)

(Formula v) :<->: (Formula v)

All [Formula v]

Some [Formula v]

None [Formula v]

ExactlyOne [Formula v]

AtMostOne [Formula v]

Let (Formula v) (Formula v -> Formula v)

Bound Integer

Orphan instances

Ord Extension

Instance details

Methods

compare :: Extension -> Extension -> Ordering

(<) :: Extension -> Extension -> Bool

(<=) :: Extension -> Extension -> Bool

(>) :: Extension -> Extension -> Bool

(>=) :: Extension -> Extension -> Bool

max :: Extension -> Extension -> Extension

min :: Extension -> Extension -> Extension

Read Extension

Instance details

Methods

readsPrec :: Int -> ReadS Extension

readList :: ReadS [Extension]

readPrec :: ReadPrec Extension

readListPrec :: ReadPrec [Extension]

NFData Extension

Instance details

Methods

rnf :: Extension -> ()

Generic a => Generic (Formula a)

Instance details

Associated Types

type Rep (Formula a) :: Type -> Type

Methods

from :: Formula a -> Rep (Formula a) x

to :: Rep (Formula a) x -> Formula a

(Generic a, NFData a) => NFData (Formula a)

Instance details

Methods

rnf :: Formula a -> ()