Language.Haskell.Tools.Refactor.Builtin.OrganizeExtensions

Documentation

highlightExtensionsQuery :: QueryChoice

organizeExtensionsRefactoring :: RefactoringChoice

projectOrganizeExtensionsRefactoring :: RefactoringChoice

projectOrganizeExtensions :: ProjectRefactoring

tryOut :: String -> String -> IO ()

organizeExtensions :: LocalRefactoring

reduceExtensions :: UnnamedModule -> Ghc [Extension]

Reduces default extension list (keeps unsupported extensions)

extensionMarkers :: UnnamedModule -> Ghc [Marker]

Collect the required extensions in a module and returns a markers associated with them

collectExtensions :: UnnamedModule -> Ghc ExtMap

Collects extensions induced by the source code (with location info)

collectExtensionsWith :: CheckNode UnnamedModule -> UnnamedModule -> Ghc ExtMap

Collects the required extensions from a module using the given traversal method

expandExtensions :: [Extension] -> [Extension]

Expands every extension in a list, while not producing any duplicates.

collectDefaultExtensions :: UnnamedModule -> [Extension]

Collects extensions enabled by default

collectTurnedOffExtensions :: UnnamedModule -> [Extension]

Collects extensions enabled by default

getExtensions :: UnnamedModule -> [String]

Collects the string representation of the extensions in the module

toExt :: String -> Maybe Extension

module Language.Haskell.Tools.Refactor.Builtin.ExtensionOrganizer.ExtMonad